generate global.json file
