+++ package.json
@@ -36,18 +36,19 @@
     "lint": "eslint . --fix",
     "format": "prettier --write .",
     "host": "npm run all && npm run preview",
-    "all": "npm run lint && npm run format && npm run check && npm run build && npm t"
+    "all": "npm run lint && npm run format && npm run check && npm run build && npm t",
+    "generate:colors": "ts-node scripts/generate-colors/index.js"
   },
   "devDependencies": {
     "@size-limit/file": "^8.2.6",
     "@sveltejs/adapter-auto": "^2.0.0",
     "@sveltejs/kit": "^1.20.4",
-    "@types/css": "^0.0.37",
     "@types/html-to-text": "^9.0.1",
     "@types/qrcode": "^1.5.1",
     "@types/rss": "^0.0.30",
     "@typescript-eslint/eslint-plugin": "^5.45.0",
     "@typescript-eslint/parser": "^5.45.0",
+    "css": "^3.0.0",
     "eslint": "^8.28.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-svelte": "^2.30.0",
@@ -66,7 +67,6 @@
     "@googleapis/calendar": "^7.0.0",
     "@js-temporal/polyfill": "^0.4.4",
     "cropperjs": "^1.5.13",
-    "css": "^3.0.0",
     "html-to-text": "^9.0.5",
     "lc-dailies": "github:acmcsufoss/lc-dailies#npm",
     "pomo": "github:ethanthatonekid/pomo#npm",
