+++ src/routes/(site)/_colors/+page.ts
@@ -1,7 +1,7 @@
 import type { PageLoadEvent } from './$types';
-import { parseGlobalCSS } from './colors';
+import type { Color } from './colors';
 
 export async function load({ fetch }: PageLoadEvent) {
-  const globalCSS = await fetch('/global.css').then((r) => r.text());
-  return { colors: parseGlobalCSS(globalCSS) };
+  const colors: Color[] = await fetch('/global.json').then((r) => r.json());
+  return { colors };
 }
